add folder refactor and hot-reload refactor
