+++ conf/webpack/Base.js
@@ -122,14 +122,14 @@ class WebpackBaseConfig {
       plugins: [],
       resolve: {
         alias: {
-          actions: `${this.srcPathAbsolute}/actions/`,
           components: `${this.srcPathAbsolute}/components/`,
           config: `${this.srcPathAbsolute}/config/${this.env}.js`,
-          images: `${this.srcPathAbsolute}/images/`,
-          sources: `${this.srcPathAbsolute}/sources/`,
-          stores: `${this.srcPathAbsolute}/stores/`,
-          srcStyles: `${this.srcPathAbsolute}/srcStyles/`,
+          containers: `${this.srcPathAbsolute}/containers/`,
           locale: `${this.srcPathAbsolute}/locale/`,
+          redux_flow: `${this.srcPathAbsolute}/redux_flow/`,
+          static: `${this.srcPathAbsolute}/static/`,
+          theme: `${this.srcPathAbsolute}/theme/`,
+          utils: `${this.srcPathAbsolute}/utils/`,
         },
         extensions: [
           '.js',
